requirements.txt
python-telegram-bot
evernote3
